Update: Edit button now leads to an EditToDoActivity.
It is still on developing mode. Need to send the intent from the selected day's schedule intent, but with the time constraint, I couldn't figure out how to. Also, I couldn't figure out how to "UPDATE" the schedule when I press the accept button in EditToDoActivity. That will be related to MainActivity.class's ResultActivity, but somehow it does not work.